+++ Leomard/AppStorage/UserPreferences.swift
@@ -29,6 +29,7 @@ final class UserPreferences: ObservableObject {
     @AppStorage("unreadonly_when_opening_inbox", store: .standard) var unreadonlyWhenOpeningInbox: Bool = true
     @AppStorage("profile_sort_method", store: .standard) var profileSortMethod: SortType = .new
     @AppStorage("check_for_update_frequency", store: .standard) var checkForUpdateFrequency: UpdateFrequency = .onceADay
+    
     // MARK: - Mark as Read Stuff
     @AppStorage("mark_post_as_read_on_open", store: .standard) var markPostAsReadOnOpen: Bool = true
     @AppStorage("mark_post_as_read_on_vote", store: .standard) var markPostAsReadOnVote: Bool = true
@@ -38,9 +39,11 @@ final class UserPreferences: ObservableObject {
     // MARK: - Blocked instances
     @AppStorage("blocked_instances", store: .standard) var blockedInstances: String = ""
     
-    // MARK: - Post View
+    // MARK: - Post
     @AppStorage("use_post_compact_view", store: .standard) var usePostCompactView: Bool = false
     @AppStorage("two_column_view", store: .standard) var twoColumnView: Bool = true
+    @AppStorage("save_liked_posts", store: .standard) var saveLikedPosts: Bool = false
+    @AppStorage("autosave_post_creation", store: .standard) var autosavePostCreation: Bool = true
     
     // MARK: - Navbar
     @AppStorage("navbar_show_letter_separators", store: .standard) var navbarShowLetterSeparators: Bool = false
@@ -50,9 +53,7 @@ final class UserPreferences: ObservableObject {
     @AppStorage("prefer_display_name_community_post", store: .standard) var preferDisplayNameCommunityPost: Bool = false
     @AppStorage("prefer_display_name_communtiy_followed", store: .standard) var preferDisplayNameCommunityFollowed: Bool = false
     @AppStorage("prefer_display_name_people_post", store: .standard) var preferDisplayNamePeoplePost: Bool = false
-    
-    // MARK: - Liked Posts
-    @AppStorage("save_liked_posts", store: .standard) var saveLikedPosts: Bool = false
+